# Conflicts:
#	Client/Source/AudioCallback.cpp
#	Client/Source/AudioCallback.h
#	Client/Source/MainComponent.cpp
#	Client/Source/MainComponent.h
#	common/JammerNetzPackage.cpp
#	common/JammerNetzPackage.h
#	third_party/JUCE-user-modules/ff_meters

+++ Client/Source/ClientConfig.cpp
@@ -53,7 +53,7 @@ void ClientConfigurator::resized()
 
 void ClientConfigurator::sliderValueChanged(Slider*)
 {
-	updateHandler_(bufferLength_.getValue(), maxLength_.getValue(), flares_.getValue());
+	updateHandler_((int) bufferLength_.getValue(), (int) maxLength_.getValue(), (int) flares_.getValue());
 }
 
 void ClientConfigurator::fromData()
